持続可能な理想社会を目指します

+++ app/api/profile/delete/route.ts
@@ -2,7 +2,6 @@
 import { NextResponse } from "next/server";
 import { createClient } from "@/lib/supabase/server";
 import { supabaseAdmin } from "@/lib/supabase/admin";
-import { Content } from "next/font/google";
 
 export async function DELETE() {
   // 認証チェック
@@ -16,6 +15,15 @@ export async function DELETE() {
   }
   const userId = user.id;
 
+  // アナウンスを作成
+  const deleteMessage = `ユーザーID ${userId}が裏切者であることが発覚したため、抹消いたしました。 市民のみなさまは安心してパラディアをご利用ください。`;
+  const { error: logError } = await supabaseAdmin
+    .from("announcements")
+    .insert({ content: deleteMessage });
+  if (logError) {
+    return NextResponse.json({ error: logError.message }, { status: 500 });
+  }
+
   // RLS バイパスして削除
   await authClient.from("likes").delete().eq("user_id", userId);
   await authClient.from("profiles").delete().eq("id", userId);
@@ -23,14 +31,6 @@ export async function DELETE() {
     await supabaseAdmin.auth.admin.deleteUser(userId);
   if (deleteError) {
     return NextResponse.json({ error: deleteError.message }, { status: 500 });
-  }
-
-  const deleteMessage = `ユーザーID ${userId}が裏切者であることが発覚したため、抹消いたしました。 市民のみなさまは安心してパラディアをご利用ください。`;
-  const { error: logError } = await authClient
-    .from("announcements")
-    .insert({ content: deleteMessage });
-  if (logError) {
-    return NextResponse.json({ error: logError.message }, { status: 500 });
   }
 
   return NextResponse.json({ success: true });
